# services/auth_service/unitTests/model/TwoFactorSetup.js
/**
 * Auth Service API
 * Authentication Service for the Transcendence Pong application. Handles JWT tokens, OAuth2/SSO, and session management. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The TwoFactorSetup model module.
 * @module model/TwoFactorSetup
 * @version 1.0.0
 */
class TwoFactorSetup {
    /**
     * Constructs a new <code>TwoFactorSetup</code>.
     * @alias module:model/TwoFactorSetup
     */
    constructor() { 
        
        TwoFactorSetup.initialize(this);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj) { 
    }

    /**
     * Constructs a <code>TwoFactorSetup</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/TwoFactorSetup} obj Optional instance to populate.
     * @return {module:model/TwoFactorSetup} The populated <code>TwoFactorSetup</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new TwoFactorSetup();

            if (data.hasOwnProperty('secret')) {
                obj['secret'] = ApiClient.convertToType(data['secret'], 'String');
            }
            if (data.hasOwnProperty('qrCodeUrl')) {
                obj['qrCodeUrl'] = ApiClient.convertToType(data['qrCodeUrl'], 'String');
            }
            if (data.hasOwnProperty('backupCodes')) {
                obj['backupCodes'] = ApiClient.convertToType(data['backupCodes'], ['String']);
            }
        }
        return obj;
    }

    /**
     * Validates the JSON data with respect to <code>TwoFactorSetup</code>.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @return {boolean} to indicate whether the JSON data is valid with respect to <code>TwoFactorSetup</code>.
     */
    static validateJSON(data) {
        // ensure the json data is a string
        if (data['secret'] && !(typeof data['secret'] === 'string' || data['secret'] instanceof String)) {
            throw new Error("Expected the field `secret` to be a primitive type in the JSON string but got " + data['secret']);
        }
        // ensure the json data is a string
        if (data['qrCodeUrl'] && !(typeof data['qrCodeUrl'] === 'string' || data['qrCodeUrl'] instanceof String)) {
            throw new Error("Expected the field `qrCodeUrl` to be a primitive type in the JSON string but got " + data['qrCodeUrl']);
        }
        // ensure the json data is an array
        if (!Array.isArray(data['backupCodes'])) {
            throw new Error("Expected the field `backupCodes` to be an array in the JSON data but got " + data['backupCodes']);
        }

        return true;
    }


}



/**
 * TOTP secret key
 * @member {String} secret
 */
TwoFactorSetup.prototype['secret'] = undefined;

/**
 * URL to QR code image for authenticator app
 * @member {String} qrCodeUrl
 */
TwoFactorSetup.prototype['qrCodeUrl'] = undefined;

/**
 * One-time backup codes
 * @member {Array.<String>} backupCodes
 */
TwoFactorSetup.prototype['backupCodes'] = undefined;






export default TwoFactorSetup;