# services/auth_service/unitTests/openApi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Auth Service API",
		"description": "Authentication Service for the Transcendence Pong application.\nHandles JWT tokens, OAuth2/SSO, and session management.\n",
		"version": "1.0.0",
		"contact": {
			"name": "Transcendence Team"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "/api/v1",
			"description": "API Gateway endpoint"
		}
	],
...
	"paths": {
...
		"/auth/2fa/enable": {
			"post": {
				"tags": [
					"Session"
				],
				"summary": "Enable 2FA",
				"description": "Generate 2FA secret and QR code for setup",
				"operationId": "enable2FA",
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "2FA setup initiated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TwoFactorSetup"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					},
					"409": {
						"description": "2FA already enabled",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"500": {
						"$ref": "#/components/responses/InternalError"
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "JWT access token"
			}
		},
		"schemas": {
...
			"TwoFactorSetup": {
				"type": "object",
				"properties": {
					"secret": {
						"type": "string",
						"description": "TOTP secret key"
					},
					"qrCodeUrl": {
						"type": "string",
						"format": "uri",
						"description": "URL to QR code image for authenticator app"
					},
					"backupCodes": {
						"type": "array",
						"items": {
							"type": "string"
						},
						"description": "One-time backup codes"
					}
				}
			},
...
			"Error": {
				"type": "object",
				"required": [
					"code",
					"message"
				],
				"properties": {
					"code": {
						"type": "string",
						"description": "Error code for client handling"
					},
					"message": {
						"type": "string",
						"description": "Human-readable error message"
					},
					"details": {
						"type": "object",
						"additionalProperties": true,
						"description": "Additional error details"
					}
				}
			}
		},
		"responses": {
...
			"Unauthorized": {
				"description": "Authentication required or token invalid",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
...
			"InternalError": {
				"description": "Internal server error",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			}
		}
	}
}